Refactor : Vo와 Dto 경계 구분이 명확하지 않아서 추가 처리. 및 in/out 구분 진행.
